+++ src/uv_fit.py
@@ -511,7 +511,7 @@ def auto_detect(vis: dict, info: dict, n_sources: int = None, clean_output=True,
     chi2 = float(np.sum(w * ((re - model.real)**2 + (im - model.imag)**2)))
     n = len(re)
     k = n_params
-    bic = k * np.log(n) + chi2
+    bic = float(k * np.log(n) + chi2)
     all_results.append({'n_sources': n_sources, 'result': result, 'bic': bic, 'chain': chain})
 
     if clean_output:
